Makefile updates, some restructure, and lint tools
